backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
argon2-cffi

